+++ boussole/conf/yaml_backend.py
@@ -1,29 +1,32 @@
 # -*- coding: utf-8 -*-
 """
-JSON settings backend
+YAML settings backend
 =====================
 """
-import json
+import yaml
+import pyaml
 
 from boussole.exceptions import SettingsBackendError
 from boussole.conf.base_backend import SettingsBackendBase
 
 
-class SettingsBackendJson(SettingsBackendBase):
+class SettingsBackendYaml(SettingsBackendBase):
     """
-    JSON backend for settings
+    YAML backend for settings
+
+    Use PyYaml for parsing and pyaml for dumping.
 
     Attributes:
         _default_filename: Filename for settings file to load.
-            Value is ``settings.json``.
+            Value is ``settings.yml``.
         _kind_name: Backend format name.
-            Value is ``json``.
+            Value is ``yaml``.
         _file_extension: Default filename extension.
-            Value is ``json``.
+            Value is ``yml``.
     """
-    _default_filename = 'settings.json'
-    _kind_name = 'json'
-    _file_extension = 'json'
+    _default_filename = 'settings.yml'
+    _kind_name = 'yaml'
+    _file_extension = 'yml'
 
     def dump(self, content, filepath, indent=4):
         """
@@ -34,11 +37,11 @@ class SettingsBackendJson(SettingsBackendBase):
             filepath (str): Settings file location.
         """
         with open(filepath, 'w') as fp:
-            json.dump(content, fp, indent=indent)
+            pyaml.dump(content, dst=fp, indent=indent)
 
     def parse(self, filepath, content):
         """
-        Parse opened settings content using JSON parser.
+        Parse opened settings content using YAML parser.
 
         Args:
             filepath (str): Settings object, depends from backend
@@ -47,15 +50,15 @@ class SettingsBackendJson(SettingsBackendBase):
 
         Raises:
             boussole.exceptions.SettingsBackendError: If parser can not decode
-                a valid JSON object.
+                a valid YAML object.
 
         Returns:
             dict: Dictionnary containing parsed setting elements.
 
         """
         try:
-            parsed = json.loads(content)
-        except ValueError:
-            msg = "No JSON object could be decoded from file: {}"
-            raise SettingsBackendError(msg.format(filepath))
+            parsed = yaml.load(content)
+        except yaml.YAMLError as exc:
+            msg = "No YAML object could be decoded from file: {}\n{}"
+            raise SettingsBackendError(msg.format(filepath, exc))
         return parsed