+++ docs/changelog.rst
@@ -8,6 +8,14 @@ Changelog
 Version 1.1.0 - Unreleased
 --------------------------
 
+* YAML backend for settings, close #7 :
+
+  * Added ``yaml_backend.SettingsBackendYaml`` backend;
+  * Implement YAML backend in unittests;
+  * Added helper to discover settings backend from filename extension;
+  * Configuration backend now implement a dump method;
+  * Changed ``project.ProjectStarter`` so it can load Configuration backend;
+
 * Don't pass anymore logger to objects, just use ``logging.getLogger("boussole")``, close #11;
 * Validate tests on Python 3.5 through tox;